\name{LBE-package}
\alias{LBE-package}
\docType{package}
\title{
Estimation of the proportion of true null hypotheses, the false discovery rate and the q-values.
}
\description{
The package LBE allows estimating the proportion of true null hypotheses and the false discovery rate. 
}
 

\details{
\tabular{ll}{
Package: \tab LBE\cr
Type: \tab Package\cr
Version: \tab 1.13.2\cr
Date: \tab 2007-1-10\cr
License: \tab GPL-2\cr
}
}
\author{
Cyril Dalmasso
Maintainer: Cyril Dalmasso \email{dalmasso@vjf.inserm.fr}
}
\references{
Dalmasso C, Broet P, Moreau T (2005): A simple procedure for estimating the false discovery rate. Bioinformatics, 21: 660 - 668.

Storey JD and Tibshirani R. (2003). Statistical significance for genome-wide studies. Proc Natl Acad Sci, 100, 9440-9445.
}
\seealso{
LBE is an alternative method to the one proposed by Storey and Tibshirani (2003) for estimating the q-values, this latter method being implemented in the package \code{\link[qvalue]{qvalue}}.
}
\keyword{ package }
